Menu System
Added a MainMenu system that allows player to load into the "Training Room" level, or quit the game. (Multiplayer scene will be added soon). Also created a Pause Menu that pauses the game when Esc is pressed and allows user to resume the game, quit the game, or load back to the main menu.