2nd Player Controller Update and Removal of Photon
Created a new player controller (named LocalPlayerController) that works with setting up a second player with their own independent move buttons. Added this buton mapping feature to the playercombat as well. Deleted Photon files from the folder as we are doing co-op instead. NOTE: GAME IS FROZEN ON PLAY ON MY SYSTEM (try on another system).